refactored common and stripped pom.xml
common was pulled apart into the sub-modules common (containing the
interfaces), and trivial_beans (containing the hello and goodbye
classes). Unit tests were added to the trivial beans, as well as a test
logback.xml.

Furthermore, the parent pom was stripped back to nothing and rebuilt
to support just common and trivial beans, and will be updated as needed
(hopefully it won't grow so large and gross).